% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_constraints.R
\name{addDistribution.Strategy}
\alias{addDistribution.Strategy}
\title{Adds distribution}
\usage{
\method{addDistribution}{Strategy}(
  this,
  component.type,
  component.label = NULL,
  variable = list(),
  label
)
}
\arguments{
\item{this}{Strategy}

\item{component.type}{character, one of rule, indicator, params}

\item{component.label}{character vector, name of component, argument 'as' is resposible for that}

\item{variable}{list, each cell should be named and this name should coincise with one of your arguments.}

\item{label}{character, name for this distribution}

\item{...}{arguments for variable}
}
\description{
Adds distribution
}
\examples{
\dontrun{
addIndicator(this, name = BBands, args = list(HLC = quote(spread), n = 20, sd = 1), as = 'bb')
paramset <- "TEST"
addDistribution(this,
    paramset.label = paramset,
    component.type = 'indicator',
    component.label = 'bb',
    variable = list(sd = seq(0.5,3,0.05)),
    label = 'bb.sd'
)

addRule(this, as = 'bb_up_dn',
     condition = (Lag(spread, 1) > Lag(bb[, 'up'], 1)) &
                  (spread < bb[,'up']) &
                  (spread > bb[,'mavg']) &
                  (abs(spread - bb[,'mavg'])  > n) ,
     type = 'enter',
     args = list(n = 0.005),
     side = -1,
     oco = 'short')
addDistribution(this,
    component.type = 'rule',
    component.label = 'bb_up_dn',
    variable = list(n = seq(0.005,0.03,0.002)),
    label = 'my_distr'
)
}
}
