% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_constraints.R
\name{addDistributionConstraint.Strategy}
\alias{addDistributionConstraint.Strategy}
\title{Adds contraints to distibutions}
\usage{
\method{addDistributionConstraint}{Strategy}(this, expr, label)
}
\arguments{
\item{this}{Strategy}

\item{expr}{expression, that contains names from labels of distributions}

\item{label}{character, name of the constraint}
}
\description{
Adds contraints to distibutions
}
\examples{
\dontrun{
addIndicator(this, name = BBands, args = list(HLC = quote(spread), n = 20, sd = 1), as = 'bb',
        lookback = 100, plot = 'main', columns = 1:3, col = "red")
addDistribution(this,
    component.type = 'indicator',
    component.label = 'bb',
    variable = list(sd = seq(0.5,3,0.05)),
    label = 'bb.sd'
)

addRule(this, as = 'bb_up_dn',
     condition = (Lag(spread,1) > Lag(bb.up,1)) &
                  (spread < bb.up) &
                  (spread > bb.mavg) &
                  (abs(spread - bb.mavg)/spread_price  > n) ,
     type = 'enter',
     args = list(n = 0.005),
     side = -1,
     oco = 'short',
     osFun = sameMoneyOs,
     osFun_args = alist(amount = 5000000))
addDistribution(this,
    component.type = 'rule',
    component.label = 'bb_up_dn',
    variable = list(n = seq(0.005,0.03,0.002)),
    label = 'my_distr'
)

addDistributionConstraint(this,
    expr = bb.sd > my_distr * 100
)
}
}
