% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericStrategy.R, R/rules_indicators.R
\name{addIndicators}
\alias{addIndicators}
\alias{addIndicator}
\alias{addIndicators.Strategy}
\title{Add multiple indicators to strategy}
\usage{
addIndicators(this, expr, name, lookback, args, lookforward, history, vars)

addIndicator(this, expr, name, lookback, args, lookforward, history, vars)

\method{addIndicators}{Strategy}(
  this,
  expr,
  name,
  lookback = NULL,
  args = list(),
  lookforward = Inf,
  history = TRUE,
  vars = NULL
)
}
\arguments{
\item{this}{model}

\item{expr}{expression}

\item{name}{character, name of signal}

\item{lookback}{numeric / expression, how many periods is needed for calculation signal}

\item{args}{list, parameters for a signal}

\item{lookforward}{numeric / expression. If it is numeric then after that number of points all indicators will be recalculated.
If it is expression then it should return logical. It will indicate recalculate indicators or not.}

\item{history}{logical, whether it expressed as matrix of previous and future values or expressed as statistic that recalculated
when lookforward triggers}

\item{vars}{character vector, which names should be exported from expr}
}
\description{
These indicators will be calculated each time when spread updates coefficients
indicators must return data.frame or matrix or array
}
