% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericStrategy.R, R/parameters.R
\name{addObject}
\alias{addObject}
\alias{addObject.Strategy}
\title{Add user-defined objects to Strategy for future usage in backtest}
\usage{
addObject(this, ...)

\method{addObject}{Strategy}(this, ...)
}
\arguments{
\item{this}{Strategy}

\item{...}{named args}
}
\description{
Add user-defined objects to Strategy for future usage in backtest
}
