% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericStrategy.R, R/rules_indicators.R
\name{addRule}
\alias{addRule}
\alias{addRule.Strategy}
\title{Add rule to strategy}
\usage{
addRule(
  this,
  expr,
  name,
  lookback,
  args,
  type,
  block,
  pathwise,
  position,
  position_const,
  price,
  on_success
)

\method{addRule}{Strategy}(
  this,
  expr = TRUE,
  name,
  lookback = 0,
  args = list(),
  type = "enter",
  block,
  pathwise = FALSE,
  position = NULL,
  position_const = NULL,
  price = NULL,
  on_success = NULL
)
}
\arguments{
\item{this}{Strategy}

\item{expr}{expression}

\item{name}{character, name of signal}

\item{lookback}{numeric / expression, how many periods is needed for calculation signal}

\item{args}{list, parameters for a signal}

\item{type}{character, 'exit' or 'enter'}

\item{block}{character, it is needed for combining bunch of rules in one namespace. For example
rule with type exit will be triggered only if it has the same block variable as entry rule.}

\item{pathwise}{logical, whether expr will be calculated on each iteration or not. In other words if pathwise is FALSE
then expr will be calculated once with indicators, it should return logical vector, then in each iteration element of this
vector will be used, if pathwise is TRUE then expr will be calculated at each iteration, it should return logical scalar.}

\item{position}{numeric vector, what position for each instrument should be at each iteration when model in position}

\item{position_const}{numeric vector, what position for each instrument should be at entry to position}

\item{price}{numeric vector, what prices should be used for change position}
}
\description{
Add rule to strategy
}
