% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules_indicators.R
\name{addRuleConstraint.Strategy}
\alias{addRuleConstraint.Strategy}
\title{Add rule's constraint}
\usage{
\method{addRuleConstraint}{Strategy}(
  this,
  expr,
  name,
  lookback = 0,
  args = list(),
  rules = NULL,
  rule_type = NULL
)
}
\arguments{
\item{this}{Strategy}

\item{expr}{expression}

\item{name}{character, name of signal}

\item{lookback}{numeric / expression, how many periods is needed for calculation signal}

\item{args}{list, parameters for a signal}

\item{rules}{character vector, names of rules}

\item{rule_type}{character, type is 'exit' or 'enter'. If it is not missed it will be applied to all rules in this group,
rules argument will be ignored.}
}
\description{
Add rule's constraint
}
