% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericStrategy.R, R/reports.R
\name{addToReport}
\alias{addToReport}
\alias{addToReport.Strategy}
\title{Add statistic to report of strategy}
\usage{
addToReport(this, ...)

\method{addToReport}{Strategy}(this, ...)
}
\arguments{
\item{this}{Strategy}

\item{...}{Stats or function objects. Function should have arguments as follows function(this, start, end), where

this - Strategy

start - numeric index of starting point

end - numeric index of ending point}
}
\description{
Add statistic to report of strategy
}
\examples{
\dontrun{
addToReport(this, Stats$sharpe,  double_sharpe=function(this, start, end){
   pnl <- getPnL(this)
   dates <- getDateByIndex(this, c(start, end))
   pnl <- pnl[paste0(dates[1], '/', dates[2])]
   sharpes <- rollapply(diff(pnl), width = 252, FUN=SharpeRatio.mean, by=30, align='right')  \%>\% na.omit
   sharpes \%>\% {mean(.) / sd(.) * sqrt(12)}
})
}

}
