% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericStrategy.R, R/timeManagement.R
\name{addTradeTime}
\alias{addTradeTime}
\alias{addTradeTime.Strategy}
\title{Adds time when trading is permitted}
\usage{
addTradeTime(this, ...)

\method{addTradeTime}{Strategy}(this, type, ...)
}
\arguments{
\item{this}{Strategy}

\item{...}{characters of start and end dates, each element is a vector of two strings that represents time of start and end}

\item{type}{character, one of enter/open (when may open position),
close (when must close position),
rebalance (when may rebalance position)}
}
\description{
Adds time when trading is permitted
}
