% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{addVariables.Strategy}
\alias{addVariables.Strategy}
\title{Add instruction for updating variables}
\usage{
\method{addVariables}{Strategy}(this, ...)
}
\arguments{
\item{this}{Strategy}

\item{...}{expression}
}
\description{
These instruction will be applied at each update of indicators and rules.
Mainly it is needed when actions like splits and dividend happen.
}
