% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustments.R
\name{adjust_close}
\alias{adjust_close}
\title{Adjust close table for dividends and splits}
\usage{
adjust_close(close, splits, dividends)
}
\arguments{
\item{close}{matrix / xts, table of close prices}

\item{splits}{matrix / xts, table of splits with default value 1}

\item{dividends}{matrix / xts, table of dividends with default value 0}
}
\value{

}
\description{
Adjust close table for dividends and splits
}
