% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtesting.R
\name{calcBacktestResults.Strategy}
\alias{calcBacktestResults.Strategy}
\title{Aggregate results from backtests}
\usage{
\method{calcBacktestResults}{Strategy}(this)
}
\arguments{
\item{this}{Strategy}
}
\description{
Aggregate results from backtests
}
