% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericStrategy.R, R/timeManagement.R
\name{clearTradeTime}
\alias{clearTradeTime}
\alias{clearTradeTime.Strategy}
\title{Remove trading time by type}
\usage{
clearTradeTime(this, type)

\method{clearTradeTime}{Strategy}(this, type = "all")
}
\arguments{
\item{this}{Strategy}

\item{type}{character, one of c('all', 'enter', 'open', 'close', 'rebalance')}
}
\description{
Remove trading time by type
}
