% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Strategy.R, R/genericStrategy.R
\name{cloneStrategy.Strategy}
\alias{cloneStrategy.Strategy}
\alias{cloneStrategy}
\title{Copy Strategy object}
\usage{
\method{cloneStrategy}{Strategy}(this, clone_data = FALSE, clone_backtest = FALSE, ...)

cloneStrategy(this, ...)
}
\arguments{
\item{this}{Strategy}

\item{clone_data}{logical, if TRUE then clone modelData object}

\item{...}{params}

\item{clone_backtests}{logical, if TRUE then all backtests will be cloned}
}
\value{
Strategy
}
\description{
Copy Strategy object
}
