% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\docType{data}
\name{corp_event}
\alias{corp_event}
\title{Indicator of corporative event}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
corp_event
}
\description{
This variable should be used in indicator's lookforward. If used then indicator will be recalculated as soon as corporative event occured.
}
\keyword{datasets}
