% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\docType{data}
\name{cur_exchanged_price}
\alias{cur_exchanged_price}
\title{Current exchanged price}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
cur_exchanged_price
}
\description{
Numeric vector.
Tt is equal to close price exchanged to data's main currency on given moment inside backtester cycle.
Often used in rule's position/position_const block.
}
\keyword{datasets}
