% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\docType{data}
\name{cur_price}
\alias{cur_price}
\title{Current price}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
cur_price
}
\description{
Numeric vector.
It is equivalent to \code{data[[data$price_table]][i, ]}.
So it is equal to close price on given moment inside backtester cycle.
Often used in rule's position/position_const block.
}
\keyword{datasets}
