% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\docType{data}
\name{data}
\alias{data}
\title{Data object.}
\format{
An object of class \code{Data} (inherits from \code{environment}) of length 27.
}
\usage{
data
}
\description{
Data. That object is passed through setData method. It contains prices and dates.
It can include such tables adjusted, close, high, low, dividends, splits and vector dates.
Often used in indicator's expr.
}
\keyword{datasets}
