% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\docType{data}
\name{dates}
\alias{dates}
\title{Vector of dates.}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
dates
}
\description{
Date vector. Vector copied from data$dates.
}
\seealso{
\code{\link{data}}
}
\keyword{datasets}
