% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Strategy.R
\name{expandLookback.Strategy}
\alias{expandLookback.Strategy}
\title{Expand lookback to the whole available data}
\usage{
\method{expandLookback}{Strategy}(this, x)
}
\arguments{
\item{this}{Strategy}

\item{x}{logical}
}
\description{
Expand lookback to the whole available data
}
