% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustments.R
\name{get_adjusted}
\alias{get_adjusted}
\title{Adjusted close from Data}
\usage{
get_adjusted(data, inds = NULL)
}
\arguments{
\item{data}{Data}

\item{inds}{numeric vector}
}
\value{

}
\description{
Adjusted close from Data
}
