% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtesting.R
\name{getBacktestPnL.Strategy}
\alias{getBacktestPnL.Strategy}
\title{Returns aggregated PnL from applyParamset backtests}
\usage{
\method{getBacktestPnL}{Strategy}(this, ind = NULL)
}
\arguments{
\item{this}{Strategy}

\item{ind}{numeric, one or multiple indexes of backtest from applyParamset}
}
\value{
xts
}
\description{
This function would work only if applyParamset was called with save.pnls = TRUE
}
