% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Strategy.R, R/genericStrategy.R
\name{getData.Strategy}
\alias{getData.Strategy}
\alias{getData}
\title{Gets modelData object from Strategy object}
\usage{
\method{getData}{Strategy}(this)

getData(this)
}
\arguments{
\item{this}{Strategy}
}
\description{
Gets modelData object from Strategy object
}
