% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Strategy.R
\name{getDateByIndex.Strategy}
\alias{getDateByIndex.Strategy}
\title{Gets dates by indexes from modelData}
\usage{
\method{getDateByIndex}{Strategy}(this, indexes = NULL)
}
\arguments{
\item{this}{Strategy}

\item{indexes}{numeric vector or NULL}
}
\value{
vector of dates
}
\description{
Gets dates by indexes from modelData
}
