% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_constraints.R
\name{getDistributions.Strategy}
\alias{getDistributions.Strategy}
\title{Get list of distributions}
\usage{
\method{getDistributions}{Strategy}(this)
}
\arguments{
\item{this}{Strategy}
}
\value{
list of distributions
}
\description{
Get list of distributions
}
