% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookbacks.R
\name{getLookback.Strategy}
\alias{getLookback.Strategy}
\title{Get lookback}
\usage{
\method{getLookback}{Strategy}(this, env = parent.frame())
}
\arguments{
\item{this}{Strategy}

\item{env}{environment}
}
\value{
numeric
}
\description{
Calculate current window for calculating all indicators and variables in model
}
