% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Strategy.R, R/genericStrategy.R
\name{getMoney.Strategy}
\alias{getMoney.Strategy}
\alias{getMoney}
\title{Gets amount of money that we have at the beginning in specific backtest}
\usage{
\method{getMoney}{Strategy}(this)

getMoney(this)
}
\arguments{
\item{this}{Strategy}

\item{from}{character, name of folder in backtests}
}
\value{
numeric, amount of money
}
\description{
Gets amount of money that we have at the beginning in specific backtest
}
