% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericStrategy.R, R/parameters.R
\name{getParams}
\alias{getParams}
\alias{getParams.Strategy}
\title{Get list of parameters according to type.}
\usage{
getParams(this, type)

\method{getParams}{Strategy}(this, type)
}
\arguments{
\item{this}{Strategy}

\item{type}{character, it can be equal to rules, indicators, pp or pm}
}
\description{
Get list of parameters according to type.
}
