% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{getPnL.Strategy}
\alias{getPnL.Strategy}
\title{Get pnl as xts series}
\usage{
\method{getPnL}{Strategy}(this, ...)
}
\arguments{
\item{this}{Strategy}

\item{...}{params}
}
\value{
getPnL.modelStrategy: xts series
}
\description{
This method should be called only after backtest is done
}
