% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{getReport.Strategy}
\alias{getReport.Strategy}
\title{Get report of Strategy}
\usage{
\method{getReport}{Strategy}(this, start, end, returns = "tibble")
}
\arguments{
\item{this}{Strategy}

\item{start}{numeric / Date / character, start of the period}

\item{end}{numeric / Date / character, end of the period}
}
\description{
Get report of Strategy
}
