% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Strategy.R, R/genericStrategy.R
\name{getSavedModels.Strategy}
\alias{getSavedModels.Strategy}
\alias{getSavedModels}
\title{Return Saved Models}
\usage{
\method{getSavedModels}{Strategy}(this)

getSavedModels(this)
}
\arguments{
\item{this}{Strategy}
}
\description{
Return Saved Models
}
