% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{getTrades.Strategy}
\alias{getTrades.Strategy}
\title{Get report of trades of Strategy}
\usage{
\method{getTrades}{Strategy}(this, start, end)
}
\arguments{
\item{this}{Strategy}

\item{start}{numeric / Date / character, start of the period}

\item{end}{numeric / Date / character, end of the period}
}
\description{
This method returns data.frame, each row is statistic for a one trade
}
