% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericStrategy.R, R/timeManagement.R
\name{getTradeTime}
\alias{getTradeTime}
\alias{getTradeTime.Strategy}
\title{Gets tarding time of strategy}
\usage{
getTradeTime(this, type)

\method{getTradeTime}{Strategy}(this, type = "all")
}
\arguments{
\item{this}{Strategy}

\item{type}{character, one of  c('all', 'enter', 'open', 'close', 'rebalance')}
}
\description{
Gets tarding time of strategy
}
