% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{getTradingLog.Strategy}
\alias{getTradingLog.Strategy}
\title{Get all transactions}
\usage{
\method{getTradingLog}{Strategy}(this, start, end)
}
\arguments{
\item{this}{Strategy}

\item{start}{numeric / Date / character, start of the period}

\item{end}{numeric / Date / character, end of the period}
}
\value{
getTradingLog: data.frame
}
\description{
Return data.frame containing columns:
Date - position change date
Rule_name - name rule
Instrument - traiding instrument
Change_pos - position change number
}
