% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Strategy.R, R/genericStrategy.R
\name{getVersion.Strategy}
\alias{getVersion.Strategy}
\alias{getVersion}
\title{Return version under which model was created}
\usage{
\method{getVersion}{Strategy}(this)

getVersion(this)
}
\arguments{
\item{this}{Strategy}
}
\value{
list
}
\description{
Return version under which model was created
}
