% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherHelpers.R
\name{Head}
\alias{Head}
\alias{Head.default}
\alias{Head.numeric}
\alias{Head.logical}
\alias{Head.character}
\alias{Head.matrix}
\title{Works as function head, erase rownames, works faster}
\usage{
Head(x, n)

\method{Head}{default}(x, n = 5L)

\method{Head}{numeric}(x, n = 5L)

\method{Head}{logical}(x, n = 5L)

\method{Head}{character}(x, n = 5L)

\method{Head}{matrix}(x, n = 5L)
}
\arguments{
\item{x}{object}

\item{n}{int}
}
\description{
Works as function head, erase rownames, works faster
}
