% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\docType{data}
\name{i}
\alias{i}
\title{Current index for main tables in data object.}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
i
}
\description{
Numeric scalar, changes over time.
Often used in rule's position.
}
\seealso{
\code{\link{data}, \link{ii}}
}
\keyword{datasets}
