% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\docType{data}
\name{ii}
\alias{ii}
\title{Current index for range based indicators.}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
ii
}
\description{
Numeric scalar, changes over time.
Often used in rule's expr when pathwise=TRUE mode or in rule's position/position_const.
}
\seealso{
\code{\link{range}, \link{i}}
}
\keyword{datasets}
