% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_constraints.R
\name{install.param.combo}
\alias{install.param.combo}
\title{Installs param combination into strategy}
\usage{
install.param.combo(strategy, param.combo)
}
\arguments{
\item{strategy}{environment, strategy environment}

\item{param.combo}{data.frame with one row}
}
\description{
Installs param combination into strategy
}
