% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_constraints.R
\name{installParams.Strategy}
\alias{installParams.Strategy}
\title{Install parameters to model}
\usage{
\method{installParams}{Strategy}(this, param.combo = NULL, ...)
}
\arguments{
\item{this}{model}

\item{param.combo}{data.frame/list/vector - parameters}

\item{...}{params}
}
\description{
Install parameters to model
}
