% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherHelpers.R
\name{Lag.matrix}
\alias{Lag.matrix}
\alias{Lag.default}
\alias{Lag}
\title{Works as function lag, but remains NA in the beginnig.}
\usage{
\method{Lag}{matrix}(x, k = 1)

\method{Lag}{default}(x, k = 1)

Lag(x, ...)
}
\arguments{
\item{x}{object}

\item{k}{numeric, integer lag}

\item{...}{params}
}
\description{
Works as function lag, but remains NA in the beginnig.
}
