% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericStrategy.R, R/lookbacks.R
\name{addLookback}
\alias{addLookback}
\alias{getLookbacks}
\alias{addLookback.Strategy}
\alias{getLookbacks.Strategy}
\title{Add lookback}
\usage{
addLookback(this, ...)

getLookbacks(this)

\method{addLookback}{Strategy}(this, ...)

\method{getLookbacks}{Strategy}(this)
}
\arguments{
\item{this}{}

\item{...}{variables}
}
\value{

}
\description{
Add dependency lookback from variable

Get variables from which whole lookback depends
}
