% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tech_analysis.R
\name{T3MA}
\alias{T3MA}
\alias{TEMA}
\alias{TRIMA}
\alias{VIDYA}
\alias{AMA}
\title{The Triple Exponential Moving Average (T3) by Tim Tillson}
\usage{
T3MA(x, n = 10, vFactor = 0.7)

TEMA(x, n = 10)

TRIMA(x, n = 10)

VIDYA(x, n)

AMA(x, n = 10, FC = 1, SC = 200, power = 1)
}
\arguments{
\item{x}{Price, volume, etc. series that is coercible to xts or matrix.}

\item{n}{numeric Number of periods to average over. Must be between 1 and nrow(x), inclusive.}

\item{FC}{numeric min period to average}

\item{SC}{numeric max period to average}

\item{power}{numeric power of adaptive weight by default it is 1, recommended value is 2}
}
\description{
The Triple Exponential Moving Average (T3) by Tim Tillson

Triple exponential moving average

Triangular Moving Average

Variable Index Dynamic Average

(Kaufmann)  Adaptive Moving Average
}
