% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\docType{data}
\name{money_init}
\alias{money_init}
\title{Initial amount of money}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
money_init
}
\description{
Numeric scalar, constant.
It is equivalent to \code{getMoney(this)}. As it is scalar it is much faster.
}
\seealso{
\code{\link{money_last}}
}
\keyword{datasets}
