% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\docType{data}
\name{money_last}
\alias{money_last}
\title{Current amount of money.}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
money_last
}
\description{
Numeric scalar, variable.
Total price of portfolio at the moment that includes cash and all bought long and sold short assets.
It can be used interchangeably with \link{money_init} if one wants to use compounded returns.
}
\seealso{
\code{\link{money_init}}
}
\keyword{datasets}
