% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{extractObjects}
\alias{extractObjects}
\alias{extractObjects.Strategy}
\title{Update environment with user-defined stats}
\usage{
extractObjects(this, e)

\method{extractObjects}{Strategy}(this, e)
}
\arguments{
\item{this}{Strategy}

\item{e}{environment}
}
\description{
Update environment with user-defined stats
}
