% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericStrategy.R
\name{perform}
\alias{perform}
\title{Simulate strategy}
\usage{
perform(
  this,
  start,
  end,
  closeAllPositionsAtTheEnd,
  index,
  install,
  calc_reports,
  ...
)
}
\arguments{
\item{this}{Strategy}

\item{start}{numeric / Date / character, index of modelData where calculation starts}

\item{end}{numeric / Date / character, index of modelData whre calculation ends}

\item{closeAllPositionsAtTheEnd}{bool, if true, then strategy closes all positions at the end of simulation}

\item{index}{numeric, index of backtesting param}

\item{install}{logical, install params in this model or create new one}

\item{calc_reports}{logical whether calculate reports right after backtest or not}

\item{...}{args}
}
\description{
Simulate strategy
}
