% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotCalendar.Strategy}
\alias{plotCalendar.Strategy}
\alias{plotCalendar.xts}
\title{Plot pnl in month-year matrix}
\usage{
\method{plotCalendar}{Strategy}(this, compounded = FALSE, ...)

\method{plotCalendar}{xts}(this, ...)
}
\arguments{
\item{this}{modelStrategy}

\item{compounded}{logical, compounded returns to use or not}

\item{...}{params}

\item{from}{character, name of backtest}
}
\description{
Plot pnl in month-year matrix
}
