% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotDrawdowns.Strategy}
\alias{plotDrawdowns.Strategy}
\title{Plot drawdowns}
\usage{
\method{plotDrawdowns}{Strategy}(this, return_type = "plot", graph_type = "ggplot2", ...)
}
\arguments{
\item{this}{Strategy}

\item{return_type}{character, plot or data}

\item{graph_type}{character, ggplot2 / xts / plotly}

\item{...}{params}

\item{from}{character, name of backtest}
}
\value{
ggplot/xts
}
\description{
Plot drawdowns
}
