% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotNetPosition}
\alias{plotNetPosition}
\alias{getNetPosition}
\alias{plotNetPosition.Strategy}
\title{Plot Net position of strategy}
\usage{
plotNetPosition(this, ...)

getNetPosition(...)

\method{plotNetPosition}{Strategy}(
  this,
  start,
  end,
  leg = "all",
  return_type = "plot",
  graph_type = "ggplot2",
  ...
)
}
\arguments{
\item{this}{Strategy}

\item{...}{params for plotNetPosition}

\item{start}{Numeric scalar / Date / character, example: start_date='2008-01-01'}

\item{end}{Numeric scalar / Date / character, example: end_date='2018-01-01'}

\item{leg}{numeric / character, numeric is responsible for capital by legs, character can be "all" then capital will be summed or it can be "sep" then
capital will be plotted for each leg}

\item{return_type}{character, enter 'plot' for graphical representation or 'data' for xts series.}

\item{graph_type}{character, ggplot2 or plotly}
}
\description{
Plot Net position of strategy

Get Capital of strategy
}
