% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotPnL.Strategy}
\alias{plotPnL.Strategy}
\title{Plot change of profit and loss through backtest period}
\usage{
\method{plotPnL}{Strategy}(
  this,
  comOn = TRUE,
  leg = "all",
  graph_type = "ggplot2",
  each_year = FALSE,
  adjust = FALSE,
  return_type = "plot",
  cutoff = FALSE,
  start,
  end
)
}
\arguments{
\item{this}{Strategy}

\item{comOn}{bool, if true then commission will be included in the 'trades' graph}

\item{leg}{numeric/character, number or numbers of legs, if it is equal to 'all' or 'sum', then all pnl among all legs
will be summed, if it is equal to 'sep', then pnl among legs will be plotted}

\item{graph_type}{character, ggplot2 / xts / plotly}

\item{each_year}{logical, if TRUE, then each graph will start with 0 each year}

\item{adjust}{logical, if TRUE, then values will be divided by getMoney(this)}

\item{return_type}{character, plot or data}

\item{cutoff}{logical, if true then vertical line will plotted on graph where model was created}
}
\description{
Plot change of profit and loss through backtest period
}
