% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotReturns.Strategy}
\alias{plotReturns.Strategy}
\title{Plot returns vs MAE/MFE}
\usage{
\method{plotReturns}{Strategy}(this, type = "MAE")
}
\arguments{
\item{this}{modelStrategy}

\item{type}{character, MAE or MFE}

\item{from}{character, name of backtest}
}
\value{
ggplot
}
\description{
Plot returns vs MAE/MFE
}
