% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericStrategy.R, R/plots.R
\name{plotReturns}
\alias{plotReturns}
\alias{plotTrades}
\alias{plotTrades.Strategy}
\title{Plot trade PnL vs MAE/MFE in trade}
\usage{
plotReturns(this, type)

plotTrades(this, type)

\method{plotTrades}{Strategy}(this, type = "MAE")
}
\arguments{
\item{this}{Strategy}

\item{type}{character, MAE or MFE}
}
\value{
ggplot
}
\description{
Red triangles is losing trades, green triangles is winning trades. Grey dashed line is y = x line.
On y-axis MAE/MFE is displayed, on x-axis PnL of trade is displayed.
}
