% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\docType{data}
\name{pos_change}
\alias{pos_change}
\title{Change of position from start of period to the end.}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
pos_change
}
\description{
Numeric vector. How many of each asset were sold or bought in that period.
It used only in commission block.
}
\keyword{datasets}
