% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\docType{data}
\name{position_last}
\alias{position_last}
\title{Size of current positions}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
position_last
}
\description{
Numeric vector.
It is equal to the number of shares bought long or sold short at the moment inside backtester cycle.
}
\keyword{datasets}
