% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericStrategy.R
\name{precalcStat}
\alias{precalcStat}
\title{Calculate statistic}
\usage{
precalcStat(this, s, start, end)
}
\arguments{
\item{this}{Strategy}

\item{s}{Stat}

\item{start}{numeric, start of the period}

\item{end}{numeric, end of the period}
}
\description{
Statistic and all dependent statistics will be calculated where possible
}
