% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericStrategy.R, R/timeManagement.R
\name{printTradeTime}
\alias{printTradeTime}
\alias{printTradeTime.Strategy}
\title{Prints trading time in format of time}
\usage{
printTradeTime(this, type)

\method{printTradeTime}{Strategy}(this, type = "all")
}
\arguments{
\item{this}{Strategy}

\item{type}{character, one of c('all', 'enter', 'open', 'close', 'rebalance')}
}
\description{
Prints trading time in format of time
}
