% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericStrategy.R, R/programParts.R
\name{addProgramPart}
\alias{addProgramPart}
\alias{getProgramParts}
\alias{getProgramPart}
\alias{addProgramPart.Strategy}
\alias{getProgramParts.Strategy}
\title{Adds variables to strategy}
\usage{
addProgramPart(this, ...)

getProgramParts(this, ...)

getProgramPart(this, part, recalc)

\method{addProgramPart}{Strategy}(this, name, ...)

\method{getProgramParts}{Strategy}(this)
}
\arguments{
\item{this}{Strategy}

\item{...}{quote / expression callbacks
There are multiple places where you can add your code :

i. full name / short name -- description.

1. init -- at initialization period before main cycle

2. after_tables / tables -- after initialization of rules tables

3. after_coefs / coefs -- after initialization of coefficients

4. each_iter / iter -- at the beginning of each iteration, before execution of rules

5. after_enter_to_pos / enter -- Right after entering to position

6. after_indicators / inds -- after initialization of indicators

7. before_enter_to_pos / bexit -- before enter to position

8. before_exit_from_pos / bexit -- before exit from position

9. after_exit_from_pos / exit -- after exit from position

10. data -- initialization of tables, derivative datasets and libraries

11. start_cycle / on_start -- at the start of cycle pass}

\item{name}{character, name of your set of program parts}
}
\description{
Adds variables to strategy

Gets list of variables(program parts)
}
