% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_classes.R
\name{RuleConstraint}
\alias{RuleConstraint}
\title{RuleConstraint constructor}
\usage{
RuleConstraint(rules = NULL, rule_type = NULL, ...)
}
\arguments{
\item{rules}{character vector, names of rules}

\item{rule_type}{character, type is 'exit' or 'enter'. If it is not missed it will be applied to all rules in this group,
rules argument will be ignored.}

\item{...}{args passed to Signal}
}
\description{
This class is needed to add constraints for expr for rules. It can be treated as operator AND for expr of rules.
}
