% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Strategy.R, R/genericStrategy.R
\name{setData.Strategy}
\alias{setData.Strategy}
\alias{setData}
\title{Set modelData object to Strategy object}
\usage{
\method{setData}{Strategy}(this, x, clearBacktest = TRUE, only = FALSE, ...)

setData(this, ...)
}
\arguments{
\item{this}{Strategy}

\item{x}{modelData}

\item{only}{logical, if it is true then method only change modelD field}

\item{...}{params}

\item{clearBacktests}{logical, if true then all backtests will be erased}
}
\description{
Set modelData object to Strategy object
}
