% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Strategy.R, R/genericStrategy.R
\name{setMoney.Strategy}
\alias{setMoney.Strategy}
\alias{setMoney}
\title{Sets amount of money, that we have at the beginning}
\usage{
\method{setMoney}{Strategy}(this, x)

setMoney(this, x)
}
\arguments{
\item{this}{Strategy}

\item{x}{numeric, amount of money}
}
\description{
Sets amount of money, that we have at the beginning
}
