% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericStrategy.R, R/parameters.R
\name{setParams}
\alias{setParams}
\alias{setParams.Strategy}
\title{Set list of parameters}
\usage{
setParams(this, ...)

\method{setParams}{Strategy}(this, args = list(), type = "all", ...)
}
\arguments{
\item{this}{Strategy}

\item{...}{args}

\item{args}{list, define your params here}

\item{type}{character, it can be equal to rules, indicators, pp, pm, com, pymodel, all}
}
\description{
Here you can define parameters for rules, indicators, program part and managers of position
}
\examples{
\dontrun{
setParams(this,
       type = 'rules',
       args = list(n = 5, a = 2)
)
}
}
