% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_classes.R
\name{Signal}
\alias{Signal}
\title{Signal constructor}
\usage{
Signal(expr, name, lookback = 0, args = list(), ...)
}
\arguments{
\item{expr}{expression}

\item{name}{character, name of signal}

\item{lookback}{numeric / expression, how many periods is needed for calculation signal}

\item{args}{list, parameters for a signal}

\item{...}{args passed to subclasses}
}
\description{
Base class for indicators, rules and more
}
