% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherHelpers.R
\name{Tail}
\alias{Tail}
\alias{Tail.default}
\alias{Tail.numeric}
\alias{Tail.logical}
\alias{Tail.character}
\alias{Tail.matrix}
\title{Works as function tail, erase rownames, works faster}
\usage{
Tail(x, n)

\method{Tail}{default}(x, n = 5L)

\method{Tail}{numeric}(x, n = 5L)

\method{Tail}{logical}(x, n = 5L)

\method{Tail}{character}(x, n = 5L)

\method{Tail}{matrix}(x, n = 5L)
}
\arguments{
\item{x}{object}

\item{n}{int}
}
\description{
Works as function tail, erase rownames, works faster
}
