% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\name{timer.Strategy}
\alias{timer.Strategy}
\title{Set up counter}
\usage{
\method{timer}{Strategy}(this, ..., env = parent.frame())
}
\arguments{
\item{this}{Strategy}

\item{...}{variables}

\item{env}{environment where variables will be defined}
}
\description{
Variables that you defined in dots will be set to 0 and will be updated by additing 1 every time updateTimer called.
}
