% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\docType{data}
\name{unreal_sum}
\alias{unreal_sum}
\title{Summed by legs unrealized pnl.}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
unreal_sum
}
\description{
Sum of profit and loss of each asset in model.
If asset has different currency from model, then pnl exchanged.
Often used in rule's expr for take and stop rules.
}
\seealso{
\code{\link{unrealized_money_last}}
}
\keyword{datasets}
