% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_constraints.R
\name{addDistribution.Strategy}
\alias{addDistribution.Strategy}
\title{Add distribution}
\usage{
\method{addDistribution}{Strategy}(
  this,
  component.type = "params",
  component.label = NULL,
  label,
  variable = list(),
  ...
)
}
\arguments{
\item{this}{Strategy}

\item{component.type}{character, one of rule, indicator, params}

\item{component.label}{character vector, name of component, argument \code{name} is resposible for that.
If component.type is \code{params} than component.label can be omitted (default value \code{all} will be passed)
or equal to argument \code{type} from \code{setParams} function}

\item{label}{character, name for this distribution. It only needed if constraints will be used.
This label will be used there as a name for this distribution.}

\item{variable}{list, This field should be used if name if one of your params coincide with arguments of this function,
if not ... can be used. Each cell must have a name and include array, as \code{list(n = c(1, 2, 4), eps = c(0.01, 0.02))}.}

\item{...}{named array of params, as \code{n = c(1, 2, 4), eps = c(0.01, 0.02)}}
}
\description{
Add distribution
}
\examples{
\dontrun{
addIndicator(this, name = 'bb', expr = {
   BBands(data$close[range, 1], n = n , sd = nsd)
 }, args = list(n = 30, nsd = 1))
addDistribution(this,
    component.type = 'indicator',
    component.label = 'bb',
    nsd = seq(0.5, 3, 0.05),
    label = 'bb.sd'
)

addRule(this, name = 'bb_up_dn',
     expr = (Lag(spread, 1) > Lag(bb[, 'up'], 1)) &
                  (spread < bb[,'up']) &
                  (spread > bb[,'mavg']) &
                  (abs(spread - bb[,'mavg'])  > n) ,
     type = 'enter',
     args = list(n = 0.005),
     block = 'short')
addDistribution(this,
    component.type = 'rule',
    component.label = 'bb_up_dn',
    n = seq(0.005,0.03,0.002),
    label = 'my_distr'
)
}
}
