% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_constraints.R
\name{addDistributionConstraint.Strategy}
\alias{addDistributionConstraint.Strategy}
\title{Add contraints to distibutions}
\usage{
\method{addDistributionConstraint}{Strategy}(this, ...)
}
\arguments{
\item{this}{Strategy}

\item{...}{unnamed expressions under distributions labels}
}
\description{
Add contraints to distibutions
}
\examples{
\dontrun{
addIndicators(this, vars = c('spread', 'bb'), expr = {
   spread <- data$mat$close[range, ] \%*\% cbind(c(0.5, -0.5))
   bb <- BBands(spread, n = n , sd = nsd)
 }, args = list(n = 30, nsd = 1))
addDistribution(this,
    component.type = 'indicator',
    component.label = 'bb',
    sd = seq(0.5,3,0.05),
    label = 'bb.sd'
)

addRule(this, name = 'bb_up_dn',
     expr = (Lag(spread, 1) > Lag(bb[, 'up'],1)) &
                  (spread < bb[, 'up']) &
                  (spread > bb[, 'mavg']) ,
     type = 'enter',
     args = list(n = 0.005),
     block = 'short',
     position_const = c(-1, 1))
addDistribution(this,
    component.type = 'rule',
    component.label = 'bb_up_dn',
    n = seq(0.005,0.03,0.002),
    label = 'my_distr'
)

addDistributionConstraint(this,
    bb.sd > my_distr * 100
)
}
}
