% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericStrategy.R, R/rules_indicators.R
\name{addIndicators}
\alias{addIndicators}
\alias{addIndicator}
\alias{addIndicators.Strategy}
\title{Add multiple indicators to strategy}
\usage{
addIndicators(this, expr, name, lookback, args, lookforward, vars)

addIndicator(this, expr, name, lookback, args, lookforward, vars)

\method{addIndicators}{Strategy}(
  this,
  expr,
  name,
  lookback = NULL,
  args = list(),
  lookforward = Inf,
  vars = NULL
)
}
\arguments{
\item{this}{model}

\item{expr}{expression}

\item{name}{character, name of signal}

\item{lookback}{numeric / expression, how many periods is needed for calculation signal}

\item{args}{list, parameters for a signal}

\item{lookforward}{numeric / expression. If it is numeric then after that number of points all indicators will be recalculated.
If it is expression then it should return logical. It will indicate recalculate indicators or not.}

\item{vars}{character vector, which names should be exported from expr}
}
\description{
Define variable via expr and name it via name. It will be recalculated each lookforward steps.
}
\details{
If it needs multiple variables it can also be done via expr, but names should be pointed in vars.
}
