% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericStrategy.R, R/reports.R
\name{calcStat}
\alias{calcStat}
\alias{calcStat.Strategy}
\title{Calculate statistic}
\usage{
calcStat(this, s, start, end, recalc, args, ...)

\method{calcStat}{Strategy}(this, s, start, end, recalc = FALSE, args = list(), ...)
}
\arguments{
\item{this}{Strategy}

\item{s}{Stat or function function(this, start, end){...}}

\item{start}{numeric, start of the period}

\item{end}{numeric, end of the period}

\item{recalc}{logical, whether stat will be recalculated with all dependent or evaluated on the basis of precalculated stats}

\item{args}{list, list of arguments for function / Stat s}

\item{...}{params to Stat if s is a function}
}
\description{
Statistic and all dependent statistics will be calculated
}
