% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commission.R, R/genericStrategy.R
\name{setCommission.Strategy}
\alias{setCommission.Strategy}
\alias{getCommission.Strategy}
\alias{setCommission}
\alias{getCommission}
\title{Set commission to model}
\usage{
\method{setCommission}{Strategy}(this, q)

\method{getCommission}{Strategy}(this)

setCommission(this, q)

getCommission(this)
}
\arguments{
\item{this}{Strategy}

\item{q}{quote/expression, it should depend from pos_change argument, but it can include any variable from model}
}
\description{
Set commission to model

Get commission expression of model
}
