% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotParamset.data.frame}
\alias{plotParamset.data.frame}
\alias{plotParamset.Strategy}
\title{Draws 5-D graph with axis x,y,size,color,symbol, which are contained in data.frame}
\usage{
\method{plotParamset}{data.frame}(
  df,
  x = NULL,
  y = NULL,
  size = NULL,
  color = NULL,
  symbol = NULL,
  size_scale = 20
)

\method{plotParamset}{Strategy}(
  this,
  x = NULL,
  y = NULL,
  size = NULL,
  color = NULL,
  symbol = NULL,
  size_scale = 20
)
}
\arguments{
\item{df}{data.frame}

\item{x}{character/expression type, axis x, default NULL}

\item{y}{character/expression type, axis y, default NULL}

\item{size}{character/expression type, axis size, default NULL}

\item{color}{character/expression type, axis color, default NULL}

\item{symbol}{character/expression type, axis symbol, default NULL}

\item{size_scale}{numeric type, point size, default 20}

\item{this}{Strategy}
}
\value{
plot_ly object
}
\description{
Draws 5-D graph with axis x,y,size,color,symbol, which are contained in data.frame

Plot backtests results in 5-D graph
}
