% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericStrategy.R, R/rules_indicators.R
\name{addIndicators}
\alias{addIndicators}
\alias{addIndicator}
\alias{addRule}
\alias{addRuleConstraint}
\alias{getIndicators}
\alias{getRules}
\alias{getRuleConstraints}
\alias{getRule}
\alias{getSignals}
\alias{removeRule}
\alias{removeIndicator}
\alias{addIndicators.Strategy}
\alias{addRule.Strategy}
\alias{addRuleConstraint.Strategy}
\alias{getIndicators.Strategy}
\alias{getRules.Strategy}
\alias{getRuleConstraints.Strategy}
\alias{getRule.Strategy}
\alias{getSignals.Strategy}
\title{Add multiple indicators to strategy}
\usage{
addIndicators(this, ...)

addIndicator(this, ...)

addRule(this, ...)

addRuleConstraint(this, ...)

getIndicators(this)

getRules(this, pathwise)

getRuleConstraints(this)

getRule(this, name)

getSignals(this)

removeRule(this, name)

removeIndicator(this, name)

\method{addIndicators}{Strategy}(this, ...)

\method{addRule}{Strategy}(this, ...)

\method{addRuleConstraint}{Strategy}(this, ...)

\method{getIndicators}{Strategy}(this)

\method{getRules}{Strategy}(this, pathwise)

\method{getRuleConstraints}{Strategy}(this)

\method{getRule}{Strategy}(this, name)

\method{getSignals}{Strategy}(this)
}
\arguments{
\item{this}{Strategy}

\item{...}{params}

\item{pathwise}{logical, if true then return only pathwise=TRUE rules}

\item{name}{character}
}
\value{
list of inforamation

list of information

list of information

list of information

list of information
}
\description{
These indicators will be calculated each time when spread updates coefficients
indicators must return data.frame or matrix or array
}
