% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericStrategy.R, R/rules_indicators.R
\name{addRuleConstraint}
\alias{addRuleConstraint}
\alias{getIndicators}
\alias{getRules}
\alias{getRuleConstraints}
\alias{getRule}
\alias{getSignals}
\alias{removeRule}
\alias{removeIndicator}
\alias{getIndicators.Strategy}
\alias{getRules.Strategy}
\alias{getRuleConstraints.Strategy}
\alias{getRule.Strategy}
\alias{getSignals.Strategy}
\alias{calcVecSignals.Strategy}
\title{Gets inforamtion about indicators in the model}
\usage{
addRuleConstraint(this, expr, name, lookback, args, rules, rule_type)

getIndicators(this)

getRules(this, pathwise)

getRuleConstraints(this)

getRule(this, name)

getSignals(this)

removeRule(this, name)

removeIndicator(this, name)

\method{getIndicators}{Strategy}(this)

\method{getRules}{Strategy}(this, pathwise)

\method{getRuleConstraints}{Strategy}(this)

\method{getRule}{Strategy}(this, name)

\method{getSignals}{Strategy}(this)

\method{calcVecSignals}{Strategy}(this, env = NULL, names = NULL, ...)
}
\arguments{
\item{this}{Strategy}

\item{name}{character}

\item{pathwise}{logical, if true then return only pathwise=TRUE rules}

\item{env}{environment, result environment, if NULL then new environment will be created}

\item{names}{character, what indicators should be evaluated}

\item{...}{variables}
}
\value{
list of inforamation

list of information

list of information

list of information

list of information

environment
}
\description{
Gets inforamtion about indicators in the model

Get information about rules in the model

Gets rule constraints

Get information about one rule in the strategy

Get indicators and rules

Calculate vector rules and indicators
}
