% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\docType{data}
\name{unrealized_money_last}
\alias{unrealized_money_last}
\title{Current unrealized profit or loss}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
unrealized_money_last
}
\description{
Numeric vector.
It is equal to profit or loss for each asset in their currencies in the model at the moment inside backtester cycle.
Often used in rule's expr block for takeprofit or stoploss rules.
}
\details{
If one want to get full unrealized pnl in main currency of model,
one should use \code{unreal_sum}. If model has only one currence it is possible to use \code{sum(unrealized_money_last)}.
}
\seealso{
\code{\link{unreal_sum}}
}
\keyword{datasets}
